# Speed up boolean indexing of flatiters