@seberg 2023-06-12
`count_boolean_trues` is explicitly vectorized/optimized.  I guess this may add a tiny bit of setup overhead for very small arrays, but I suspect it doesn't matter and I also really like reusing the helper :).

Thanks @Exaphis!  In principle we could also specialize the `copyswap` for trivial copies.  Although I am generally trying to remove it right now to future-proof for new DTypes. `npy_fasttake_impl` is an example of that pattern.

(That could speed up things quite a bit except for `object` dtype.)